controller/requirements.txt
cbor==1.0.0
coppeliasim_zmqremoteapi_client==2.0.4
numpy==2.3.1
paho-mqtt==2.1.0
pyzmq==27.0.0
pandas
openpyxl
